refactor: delete dead instrumentedAction() wrapper
The wrapper was exported but only its tests referenced it. The actual
track() site is the program-level postAction hook in index.ts.
classifyError() in the same file IS still in use; kept and the file
renamed from lib/instrumented-action.ts to lib/classify-error.ts.
